+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.9.1 - TBD
+- Add `am-now` directive to show current time ([#119](https://github.com/urish/angular-moment/pull/119), contributed by [bbodenmiller](https://github.com/bbodenmiller)
 - Add support for locale strings customization ([#102](https://github.com/urish/angular-moment/pull/102), contributed by [vosi](https://github.com/vosi))
 - Support for changing the timezone via `amMoment.changeTimezone()` ([#92](https://github.com/urish/angular-moment/issues/92))
 - Support for AngularJS 1.4.x